refactor(lib): extract analyze_symbol and analyze_module into tools/ (#1189)
